+++ examples/networking/sdn/main.tf
@@ -1,0 +1,2 @@
+data "vkcs_networking_sdn" "sdn" {
+}